client/pyproject.toml
[project]
name = "agent-registry-client"
version = "0.1.0"
description = "Python client SDK for Agent Registry API"
authors = [
    {name = "Agent Registry Team", email = "team@example.com"},
]
dependencies = [
    "boto3>=1.34.0",
    "a2a-sdk>=0.1.0",
    "requests>=2.31.0",
    "requests-aws4auth>=1.2.3",
]
requires-python = ">=3.11"
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "moto>=4.2.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "moto>=4.2.0",
]
lint = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing"

[dependency-groups]
dev = [
    "moto>=5.1.12",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
